+++ saleor/csv/notifications.py
@@ -40,6 +40,8 @@ def send_export_download_link_notification(export_file: "ExportFile", data_type:
     manager.notify(NotifyEventType.CSV_EXPORT_SUCCESS, payload)
     if data_type == "gift cards":
         manager.gift_card_export_completed(export_file)
+    if data_type == "products":
+        manager.product_export_completed(export_file)
 
 
 def send_export_failed_info(export_file: "ExportFile", data_type: str):
@@ -54,3 +56,5 @@ def send_export_failed_info(export_file: "ExportFile", data_type: str):
     manager.notify(NotifyEventType.CSV_EXPORT_FAILED, payload)
     if data_type == "gift cards":
         manager.gift_card_export_completed(export_file)
+    if data_type == "products":
+        manager.product_export_completed(export_file)