+++ saleor/plugins/webhook/tests/subscription_webhooks/test_create_deliveries_for_subscription.py
@@ -26,8 +26,8 @@ from .payloads import (
     generate_category_payload,
     generate_collection_payload,
     generate_customer_payload,
+    generate_export_payload,
     generate_fulfillment_payload,
-    generate_gift_card_export_payload,
     generate_gift_card_payload,
     generate_invoice_payload,
     generate_menu_item_payload,
@@ -795,7 +795,7 @@ def test_gift_card_export_completed(
     )
 
     # then
-    expected_payload = generate_gift_card_export_payload(user_export_file, gift_card_id)
+    expected_payload = generate_export_payload(user_export_file, gift_card_id)
     assert deliveries[0].payload.payload == expected_payload
     assert len(deliveries) == len(webhooks)
     assert deliveries[0].webhook == webhooks[0]
@@ -1225,6 +1225,32 @@ def test_product_metadata_updated(
     deliveries = create_deliveries_for_subscriptions(event_type, product, webhooks)
     expected_payload = json.dumps({"product": {"id": product_id}})
 
+    assert deliveries[0].payload.payload == expected_payload
+    assert len(deliveries) == len(webhooks)
+    assert deliveries[0].webhook == webhooks[0]
+
+
+def test_product_export_completed(
+    user_export_file, tmpdir, subscription_product_export_completed_webhook
+):
+    # given
+    file_mock = MagicMock(spec=File)
+    file_mock.name = "temp_file.csv"
+
+    user_export_file.content_file = file_mock
+    user_export_file.save()
+
+    webhooks = [subscription_product_export_completed_webhook]
+    event_type = WebhookEventAsyncType.PRODUCT_EXPORT_COMPLETED
+    export_id = graphene.Node.to_global_id("ExportFile", user_export_file.id)
+
+    # when
+    deliveries = create_deliveries_for_subscriptions(
+        event_type, user_export_file, webhooks
+    )
+
+    # then
+    expected_payload = generate_export_payload(user_export_file, export_id)
     assert deliveries[0].payload.payload == expected_payload
     assert len(deliveries) == len(webhooks)
     assert deliveries[0].webhook == webhooks[0]